chore: extract risk options
